Add function for authenticating user
