+++ main.py
@@ -5,8 +5,8 @@ from fastapi.staticfiles import StaticFiles
 from pydantic import BaseModel, EmailStr, Field, HttpUrl 
 from sqlalchemy.exc import IntegrityError
 from sqlmodel import select
-from database import SessionDep, ShortenedURL, User 
-from pwdlib import PasswordHash
+from database import SessionDep, ShortenedURL, User
+from security import get_hashed_password 
 
 CODE_MAX_RETRY = 10
 
@@ -30,14 +30,6 @@ class UserPublic(BaseModel):
     id: int
     email: str
     disabled: bool
-
-password_hash = PasswordHash.recommended()
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return password_hash.verify(plain_password, hashed_password)
-
-def get_hashed_password(plain_password: str) -> str:
-    return password_hash.hash(plain_password)
 
 @app.exception_handler(status.HTTP_404_NOT_FOUND)
 def http_not_found_exception_handler(request, exc):
